requirements.txt
Flask==1.1.2
Jinja2==2.11.2
requests==2.9.1
Werkzeug==1.0.1
markdown
wtforms
flask_wtf
flask_login
itsdangerous
datetime
flask_sqlalchemy
pymysql
flask_bcrypt
flask_mail
Pillow
email_validator
